moved camera script
