+++ src/cs/GameLoop.cs
@@ -59,13 +59,6 @@ public partial class GameLoop : Node2D {
 	private int RemainingTurns; // The number of turns remaining until the end of the game
 
 	private ResourceManager RM;
-	
-	private Camera2D Camera;
-	private Vector2 ZOOM_MIN = new Vector2(0.5f,0.5f);
-	private Vector2 ZOOM_MAX = new Vector2(1f,1f);
-	private Vector2 ZOOM_SPEED = new Vector2(0.2f,0.2f);
-	private Vector2 ZoomVal = new Vector2(1.0f,1.0f);
-	
 
 	//TODO: Add Shocks once they are implemented
 
@@ -103,8 +96,6 @@ public partial class GameLoop : Node2D {
 		
 		// Fetch resource manager
 		RM = GetNode<ResourceManager>("ResourceManager");
-		
-		Camera = GetNode<Camera2D>("World/Camera2D");
 
 		// Initially set all plants form their configs
 		foreach(PowerPlant pp in PowerPlants) {
@@ -267,30 +258,6 @@ public partial class GameLoop : Node2D {
 	public void _OnNextTurn() {
 		if(GS == GameState.PLAYING) {
 			NewTurn();
-		}
-	}
-	
-	// Takes Input events that are not handled by nodes or controls
-	public override void _UnhandledInput(InputEvent @event) {
-	// Camera can be moved by holding left click and dragging the mouse
-		if (@event is InputEventMouseMotion MouseMotion) {
-			if (MouseMotion.ButtonMask == MouseButtonMask.Left)
-				Camera.Position -= MouseMotion.Relative / Camera.Zoom;
-		}
-	// Can zoom the camera using the mouse wheel, smoothed with a tween animation
-		if (@event is InputEventMouseButton MouseBtn) {
-			if (MouseBtn.ButtonIndex == MouseButton.WheelDown)
-				if (Camera.Zoom > ZOOM_MIN) {
-					ZoomVal = Camera.Zoom - ZOOM_SPEED;
-					Tween TweenZoomIn = CreateTween();
-					TweenZoomIn.TweenProperty(Camera, "zoom", ZoomVal, 0.3f);
-				}
-			if (MouseBtn.ButtonIndex == MouseButton.WheelUp)
-				if (Camera.Zoom < ZOOM_MAX) {
-					ZoomVal = Camera.Zoom + ZOOM_SPEED;
-					Tween TweenZoomOut = CreateTween();
-					TweenZoomOut.TweenProperty(Camera, "zoom", ZoomVal, 0.3f);
-				}
 		}
 	}
 }
